Split hips/draw/simple.py in two files paint.py and ui.py
