+++ hips/draw/__init__.py
@@ -1,3 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """HiPS tile drawing."""
-from .simple import *
+from .paint import *
+from .ui import *
